Remvoed pages/index.html

+++ template.sh
@@ -5,11 +5,12 @@ top=$(<top.html)
 bottom=$(<bottom.html)
 for f in *.html 
 do
-    echo "$f"
-    rm pages/$f
-    echo "$top" >> pages/$f
-    cat $f >> pages/$f
-    echo "$bottom" >> pages/$f
+    if [ "$f" != "index.html" ]; then
+        echo "$f"
+        rm pages/$f
+        echo "$top" >> pages/$f
+        cat $f >> pages/$f
+        echo "$bottom" >> pages/$f
+    fi
 done
 echo "saved above files in pages/"
-
